+++ internal/doc/main.go
@@ -61,7 +61,8 @@ import (
 	openshift4_19 "github.com/coreos/butane/config/openshift/v4_19"
 	openshift4_20 "github.com/coreos/butane/config/openshift/v4_20"
 	openshift4_21 "github.com/coreos/butane/config/openshift/v4_21"
-	openshift4_22_exp "github.com/coreos/butane/config/openshift/v4_22_exp"
+	openshift4_22 "github.com/coreos/butane/config/openshift/v4_22"
+	openshift4_23_exp "github.com/coreos/butane/config/openshift/v4_23_exp"
 	openshift4_8 "github.com/coreos/butane/config/openshift/v4_8"
 	openshift4_9 "github.com/coreos/butane/config/openshift/v4_9"
 	r4e1_0 "github.com/coreos/butane/config/r4e/v1_0"
@@ -134,7 +135,7 @@ func generate(dir string) error {
 			"openshift",
 			[]version{
 				// inverse order of website navbar
-				{"4.22.0-experimental", openshift4_22_exp.Config{}},
+				{"4.23.0-experimental", openshift4_23_exp.Config{}},
 				{"4.8.0", openshift4_8.Config{}},
 				{"4.9.0", openshift4_9.Config{}},
 				{"4.10.0", openshift4_10.Config{}},
@@ -149,6 +150,7 @@ func generate(dir string) error {
 				{"4.19.0", openshift4_19.Config{}},
 				{"4.20.0", openshift4_20.Config{}},
 				{"4.21.0", openshift4_21.Config{}},
+				{"4.22.0", openshift4_22.Config{}},
 			},
 		},
 		{