+++ docs/release-notes.md
@@ -10,9 +10,12 @@ nav_order: 9
 
 ### Features
 
+- Stabilize OpenShift spec 4.22.0, targeting Ignition spec 3.6.0
+- Add OpenShift spec 4.23.0-experimental, targeting Ignition spec
+  3.7.0-experimental
 - Add `systemd.quadlets` section for embedding Podman Quadlet files
   _(fcos 1.8.0-exp, fiot 1.1.0-exp, flatcar 1.2.0-exp, openshift
-  4.22.0-exp, r4e 1.2.0-exp)_
+  4.22.0+, r4e 1.2.0-exp)_
 
 ### Bug fixes